Added animation and cube rotate
Now you can rotate a face of the cube with mouse.
You can select a threshold to rotate or not the face (if
it is near to 0 degrees, it will not rotate).

There is a graphical bug. Sometimes, when you rotate a face,
a face element rotates in a incorrect way.